Refactor SDK classes into separate modules
